refactor(app): move the pane accessors into their own module
The window exposes a separate property per result pane (`cells` /
`p1_cells` and so on) instead of an indexed model, so 44 tiny wrappers
exist purely to fork on `pane == 0`. They were scattered through
main.rs between unrelated logic.

Collecting them in `pane` puts the fork in one place and takes 340 lines
of boilerplate out of the way of the code that actually does something.
Call sites are unchanged — main.rs glob-imports the module.
