+++ app/src/main.rs
@@ -21,6 +21,7 @@ mod export;
 mod format;
 mod mock;
 mod model;
+mod pane;
 mod query_parse;
 mod self_update;
 #[cfg(feature = "mock")]
@@ -36,6 +37,7 @@ use std::sync::{Arc, Mutex};
 use slint::{Model, ModelRc, SharedString, VecModel};
 
 use dispatch::AnyDriver;
+use pane::*;
 
 /// Label used for connections with no explicit group.
 const UNGROUPED: &str = "Ungrouped";
@@ -2858,101 +2860,16 @@ fn clear_grid(w: &MainWindow, pane: usize) {
     );
 }
 
-// ----- per-pane result setters: pane 0 writes the base properties, pane 1 the
-// p1-* mirror. Workspace groups render independently, including table chrome.
-fn set_p_cells(w: &MainWindow, pane: usize, m: ModelRc<GridCell>) {
-    if pane == 0 {
-        w.set_grid_cells(m);
-    } else {
-        w.set_p1_cells(m);
-    }
-}
-fn set_p_columns(w: &MainWindow, pane: usize, m: ModelRc<GridColumn>) {
-    if pane == 0 {
-        w.set_grid_columns(m);
-    } else {
-        w.set_p1_columns(m);
-    }
-}
-fn set_p_col_count(w: &MainWindow, pane: usize, n: i32) {
-    if pane == 0 {
-        w.set_grid_col_count(n);
-    } else {
-        w.set_p1_col_count(n);
-    }
-}
-fn set_p_col_widths(w: &MainWindow, pane: usize, m: ModelRc<f32>) {
-    if pane == 0 {
-        w.set_grid_col_widths(m);
-    } else {
-        w.set_p1_col_widths(m);
-    }
-}
-fn set_p_result_kind(w: &MainWindow, pane: usize, k: i32) {
-    if pane == 0 {
-        w.set_result_kind(k);
-    } else {
-        w.set_p1_result_kind(k);
-    }
-}
-fn set_p_result_status(w: &MainWindow, pane: usize, s: SharedString) {
-    if pane == 0 {
-        w.set_result_status(s);
-    } else {
-        w.set_p1_result_status(s);
-    }
-}
-fn set_p_status_error(w: &MainWindow, pane: usize, b: bool) {
-    if pane == 0 {
-        w.set_status_error(b);
-    } else {
-        w.set_p1_status_error(b);
-    }
-}
-fn set_p_active_table(w: &MainWindow, pane: usize, table: SharedString) {
-    if pane == 0 {
-        w.set_active_table(table);
-    } else {
-        w.set_p1_active_table(table);
-    }
-}
-fn set_p_total_rows(w: &MainWindow, pane: usize, total: i32) {
-    if pane == 0 {
-        w.set_total_rows(total);
-    } else {
-        w.set_p1_total_rows(total);
-    }
-}
-fn set_p_page_bounds(w: &MainWindow, pane: usize, start: i32, end: i32, prev: bool, next: bool) {
-    if pane == 0 {
-        w.set_page_start(start);
-        w.set_page_end(end);
-        w.set_can_prev(prev);
-        w.set_can_next(next);
-    } else {
-        w.set_p1_page_start(start);
-        w.set_p1_page_end(end);
-        w.set_p1_can_prev(prev);
-        w.set_p1_can_next(next);
-    }
-}
-fn set_p_read_only(w: &MainWindow, pane: usize, read_only: bool) {
-    if pane == 0 {
-        w.set_grid_read_only(read_only);
-    } else {
-        w.set_p1_grid_read_only(read_only);
-    }
-}
 /// The editor's error highlight, all 0-based in the full buffer: the line the
 /// engine pointed at, the span of the statement it belongs to, and the token to
 /// underline (`len == 0` when the engine reported no position, only a line).
 #[derive(Clone, Copy)]
-struct ErrorMark {
-    line: i32,
-    from: i32,
-    to: i32,
-    col: i32,
-    len: i32,
+pub(crate) struct ErrorMark {
+    pub(crate) line: i32,
+    pub(crate) from: i32,
+    pub(crate) to: i32,
+    pub(crate) col: i32,
+    pub(crate) len: i32,
 }
 
 /// Shift a driver's error spot into buffer coordinates. `origin` is where the
@@ -2968,125 +2885,6 @@ fn mark_from(spot: &editor::ErrorSpot, origin: (i32, i32)) -> ErrorMark {
     }
 }
 
-/// Arm (or clear) the editor's error highlight for `pane`. `None` clears it.
-fn set_p_error_mark(w: &MainWindow, pane: usize, mark: Option<ErrorMark>) {
-    // One gate for every caller: the mark stays in pane state either way, so
-    // turning the preference back on can re-arm it without a re-run.
-    let mark = mark.filter(|_| w.get_error_highlight());
-    let m = mark.unwrap_or(ErrorMark {
-        line: -1,
-        from: -1,
-        to: -1,
-        col: 0,
-        len: 0,
-    });
-    if pane == 0 {
-        w.set_error_line(m.line);
-        w.set_error_from(m.from);
-        w.set_error_to(m.to);
-        w.set_error_col(m.col);
-        w.set_error_len(m.len);
-    } else {
-        w.set_p1_error_line(m.line);
-        w.set_p1_error_from(m.from);
-        w.set_p1_error_to(m.to);
-        w.set_p1_error_col(m.col);
-        w.set_p1_error_len(m.len);
-    }
-}
-fn set_p_pending_count(w: &MainWindow, pane: usize, n: i32) {
-    if pane == 0 {
-        w.set_pending_count(n);
-    } else {
-        w.set_p1_pending_count(n);
-    }
-}
-fn set_p_index_rows(w: &MainWindow, pane: usize, rows: ModelRc<IndexRow>) {
-    if pane == 0 {
-        w.set_index_rows(rows);
-    } else {
-        w.set_p1_index_rows(rows);
-    }
-}
-fn set_p_results_meta(w: &MainWindow, pane: usize, s: SharedString) {
-    if pane == 0 {
-        w.set_results_meta(s);
-    } else {
-        w.set_p1_results_meta(s);
-    }
-}
-fn set_p_chart_bars(w: &MainWindow, pane: usize, m: ModelRc<ChartBar>) {
-    if pane == 0 {
-        w.set_chart_bars(m);
-    } else {
-        w.set_p1_chart_bars(m);
-    }
-}
-fn set_p_grid_sort(w: &MainWindow, pane: usize, col: i32, asc: bool) {
-    if pane == 0 {
-        w.set_grid_sort_col(col);
-        w.set_grid_sort_asc(asc);
-    } else {
-        w.set_p1_grid_sort_col(col);
-        w.set_p1_grid_sort_asc(asc);
-    }
-}
-fn set_p_grid_col_filters(w: &MainWindow, pane: usize, m: ModelRc<SharedString>) {
-    if pane == 0 {
-        w.set_grid_col_filters(m);
-    } else {
-        w.set_p1_grid_col_filters(m);
-    }
-}
-fn set_p_filter_columns(w: &MainWindow, pane: usize, m: ModelRc<SharedString>) {
-    if pane == 0 {
-        w.set_filter_columns(m);
-    } else {
-        w.set_p1_filter_columns(m);
-    }
-}
-fn set_p_filter_col(w: &MainWindow, pane: usize, v: SharedString) {
-    if pane == 0 {
-        w.set_filter_col(v);
-    } else {
-        w.set_p1_filter_col(v);
-    }
-}
-fn set_p_grid_filter(w: &MainWindow, pane: usize, v: SharedString) {
-    if pane == 0 {
-        w.set_grid_filter(v);
-    } else {
-        w.set_p1_grid_filter(v);
-    }
-}
-fn set_p_detail_pretty(w: &MainWindow, pane: usize, m: ModelRc<SharedString>) {
-    if pane == 0 {
-        w.set_grid_detail_pretty(m);
-    } else {
-        w.set_p1_detail_pretty(m);
-    }
-}
-fn set_p_selected_row(w: &MainWindow, pane: usize, row: i32) {
-    if pane == 0 {
-        w.set_selected_row(row);
-    } else {
-        w.set_p1_selected_row(row);
-    }
-}
-fn set_p_range_anchor(w: &MainWindow, pane: usize, row: i32) {
-    if pane == 0 {
-        w.set_range_anchor_row(row);
-    } else {
-        w.set_p1_range_anchor_row(row);
-    }
-}
-fn set_p_range_anchor_col(w: &MainWindow, pane: usize, col: i32) {
-    if pane == 0 {
-        w.set_range_anchor_col(col);
-    } else {
-        w.set_p1_range_anchor_col(col);
-    }
-}
 /// Slice `grid` down to the pane's finalized drag/shift-click range, if any
 /// is active and still in bounds; otherwise return the grid unchanged. A
 /// range is always exactly one column wide (the column the drag started
@@ -3129,13 +2927,6 @@ fn detail_pretty_row(g: &model::GridModel, row: usize) -> Vec<SharedString> {
 fn refresh_detail_pretty(w: &MainWindow, pane: usize, g: &model::GridModel, row: i32) {
     let rows = detail_pretty_row(g, row.max(0) as usize);
     set_p_detail_pretty(w, pane, ModelRc::from(Rc::new(VecModel::from(rows))));
-}
-fn get_p_col_widths(w: &MainWindow, pane: usize) -> Vec<f32> {
-    if pane == 0 {
-        w.get_grid_col_widths().iter().collect()
-    } else {
-        w.get_p1_col_widths().iter().collect()
-    }
 }
 /// Snapshot the live client-side grid view of a group, to stash on the result the
 /// user is leaving so switching back restores filters/sort/hidden/order/widths.
@@ -3474,141 +3265,6 @@ fn build_palette_items(
     }
     (items, actions)
 }
-fn set_p_editing(w: &MainWindow, pane: usize, row: i32, col: i32) {
-    if pane == 0 {
-        w.set_editing_row(row);
-        w.set_editing_col(col);
-    } else {
-        w.set_p1_editing_row(row);
-        w.set_p1_editing_col(col);
-    }
-}
-fn set_p_editing_large(w: &MainWindow, pane: usize, large: bool) {
-    if pane == 0 {
-        w.set_editing_large(large);
-    } else {
-        w.set_p1_editing_large(large);
-    }
-}
-fn set_p_doc_tree(w: &MainWindow, pane: usize, m: ModelRc<DocRow>) {
-    if pane == 0 {
-        w.set_doc_tree(m);
-    } else {
-        w.set_p1_doc_tree(m);
-    }
-}
-fn set_p_query_running(w: &MainWindow, pane: usize, b: bool) {
-    if pane == 0 {
-        w.set_query_running(b);
-    } else {
-        w.set_p1_query_running(b);
-    }
-}
-fn set_p_streaming(w: &MainWindow, pane: usize, b: bool) {
-    if pane == 0 {
-        w.set_streaming(b);
-    } else {
-        w.set_p1_streaming(b);
-    }
-}
-fn set_p_completion_items(w: &MainWindow, pane: usize, m: ModelRc<PaletteItem>) {
-    if pane == 0 {
-        w.set_completion_items(m);
-    } else {
-        w.set_p1_completion_items(m);
-    }
-}
-fn set_p_completion_visible(w: &MainWindow, pane: usize, b: bool) {
-    if pane == 0 {
-        w.set_completion_visible(b);
-    } else {
-        w.set_p1_completion_visible(b);
-    }
-}
-fn get_p_completion_visible(w: &MainWindow, pane: usize) -> bool {
-    if pane == 0 {
-        w.get_completion_visible()
-    } else {
-        w.get_p1_completion_visible()
-    }
-}
-fn set_p_completion_selected(w: &MainWindow, pane: usize, i: i32) {
-    if pane == 0 {
-        w.set_completion_selected(i);
-    } else {
-        w.set_p1_completion_selected(i);
-    }
-}
-fn get_p_completion_selected(w: &MainWindow, pane: usize) -> i32 {
-    if pane == 0 {
-        w.get_completion_selected()
-    } else {
-        w.get_p1_completion_selected()
-    }
-}
-fn p_completion_count(w: &MainWindow, pane: usize) -> i32 {
-    if pane == 0 {
-        w.get_completion_items().row_count() as i32
-    } else {
-        w.get_p1_completion_items().row_count() as i32
-    }
-}
-fn set_p_find_open(w: &MainWindow, pane: usize, b: bool) {
-    if pane == 0 {
-        w.set_find_open(b);
-    } else {
-        w.set_p1_find_open(b);
-    }
-}
-fn get_p_find_open(w: &MainWindow, pane: usize) -> bool {
-    if pane == 0 {
-        w.get_find_open()
-    } else {
-        w.get_p1_find_open()
-    }
-}
-fn set_p_find_text(w: &MainWindow, pane: usize, s: SharedString) {
-    if pane == 0 {
-        w.set_find_text(s);
-    } else {
-        w.set_p1_find_text(s);
-    }
-}
-fn get_p_find_text(w: &MainWindow, pane: usize) -> String {
-    if pane == 0 {
-        w.get_find_text().to_string()
-    } else {
-        w.get_p1_find_text().to_string()
-    }
-}
-fn set_p_find_status(w: &MainWindow, pane: usize, s: SharedString) {
-    if pane == 0 {
-        w.set_find_status(s);
-    } else {
-        w.set_p1_find_status(s);
-    }
-}
-/// Nudge the editor to scroll the cursor into view (e.g. after a find jump).
-/// A plain counter bump, not a bound value — see `scroll-request` in
-/// code-editor.slint for why it has to be one-shot.
-fn bump_p_scroll_request(w: &MainWindow, pane: usize) {
-    if pane == 0 {
-        w.set_scroll_request(w.get_scroll_request().wrapping_add(1));
-    } else {
-        w.set_p1_scroll_request(w.get_p1_scroll_request().wrapping_add(1));
-    }
-}
-fn get_p_cursor(w: &MainWindow, pane: usize) -> (usize, usize) {
-    if pane == 0 {
-        (w.get_cursor_line() as usize, w.get_cursor_col() as usize)
-    } else {
-        (
-            w.get_p1_cursor_line() as usize,
-            w.get_p1_cursor_col() as usize,
-        )
-    }
-}
-
 /// Does any cell of `row` contain `needle` (already lowercased)? An empty
 /// needle matches every row.
 fn row_contains(row: &[model::VmCell], needle: &str) -> bool {
